feat: recent channel widget
